implement a generic solution for setting a react key on a component from outside the component.
This has the advantage that neither every component that may be part of a sequence has to
implement a :keyfn function nor does the component have to know that fact and how to generate
a key which will not be needed by the component for any other purpose.

To keep the current API and since the key actually IS metadata this is accomplished
by setting metadata on the components value at call time using the with-key
function, e.g. (MyComponent (with-key value key) ...), which will override :keyfn.

The only restriction for this to work is that the component's value has to be a datatype
which may carry metadata (implements the IWithMeta interface) like vectors, hashmaps, etc...